Make hex numbers always unsigned

+++ print_numbers.c
@@ -6,7 +6,7 @@
 /*   By: vfurmane <vfurmane@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/17 09:56:41 by vfurmane          #+#    #+#             */
-/*   Updated: 2021/01/20 11:23:47 by vfurmane         ###   ########.fr       */
+/*   Updated: 2021/01/20 11:45:13 by vfurmane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,15 +59,15 @@ char	*ft_print_int(va_list args, long int precision, int isunsigned)
 
 char	*ft_print_hex(va_list args, int lower)
 {
-	int			len;
-	char		*set;
-	char		*str;
-	long int	nb;
+	int				len;
+	char			*set;
+	char			*str;
+	unsigned int	nb;
 
 	set = "0123456789ABCDEF";
 	if (lower == 1)
 		set = "0123456789abcdef";
-	nb = (long int)va_arg(args, int);
+	nb = va_arg(args, unsigned int);
 	if (nb < 0)
 		nb = -nb;
 	len = ft_count_digits_hex(nb);
